% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDDconf_epiverse-package.R
\docType{package}
\name{IDDconf_epiverse-package}
\alias{IDDconf_epiverse}
\alias{IDDconf_epiverse-package}
\title{IDDconf_epiverse: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mbkoltai/IDDconf_epiverse}
  \item Report bugs at \url{https://github.com/mbkoltai/IDDconf_epiverse/issues}
}

}
\author{
\strong{Maintainer}: Mihaly Koltai \email{kolmisi@gmail.com} [copyright holder]

}
\keyword{internal}
